Remove dependency on modified k8s.io/apiserver
Previously, using this repo required using a modified version
of k8s.io/apiserver.  This commit removes that dependency, opting
instead to include a stripped-down copy of the API server installer
that specifically serves the needs of custom metrics API servers.

This also fixes a bug in which certain endpoints were not always
returning MetricValueLists.