Remove references to the sample adapter
The sample adapter was replaced with the basic adapter which we use for
writing e2e tests in Kubernetes (since it's not a whole lot more
complicated).  There were still references to the old sample adapter in
the Makefile, README, and getting started guide, so this removes/updates
those references.