pkg/cmd: allow users to configure audit logs
Audit logs can be very helpful when debugging request coming to the
metrics server. They provide insights on the requested resources as well
are information on the user that sent the requests. This is helpful when
trying to narrow down applications spamming the server. Also, we can
configure the audit logs to provide even more information such as the
content of the response which is very helpful when trying to introspect
incorrect data.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>